+++ src/UiPopover.vue
@@ -8,18 +8,18 @@
         :contain-focus="containFocus"
         :focus-redirector="focusRedirector"
 
-        @focus-overflow="close"
-        @keydown.native.esc="close"
+        @focus-overflow="close()"
     >
         <slot></slot>
     </ui-focus-container>
 </template>
 
 <script>
-import tippy from 'tippy.js/dist/esm/tippy.js';
+import tippy from 'tippy.js/esm';
 
 import classlist from './helpers/classlist';
 import elementRef from './helpers/element-ref';
+import events from './helpers/events';
 import UiFocusContainer from './UiFocusContainer.vue';
 
 export default {
@@ -28,7 +28,7 @@ export default {
     props: {
         animation: {
             type: String,
-            default: 'fade' // 'fade', 'shift-away', or 'none'
+            default: 'fade' // 'fade', 'shift-away', 'scale', or 'none'
         },
         appendToBody: {
             type: Boolean,
@@ -51,13 +51,9 @@ export default {
             default: false
         },
         focusRedirector: Function,
-        offset: {
-            type: [Number, String],
-            default: 0
-        },
         openOn: {
             type: String,
-            default: 'click' // 'click', 'mouseenter', 'focus', or 'always'
+            default: 'click' // 'click', 'mouseenter', 'focus', or 'manual', plus 'hover' (compat)
         },
         position: {
             type: String,
@@ -95,9 +91,11 @@ export default {
         }
     },
 
-    mounted() {
+    created() {
         this.tip = null;
-        this.lastFocusedElement = null;
+    },
+
+    mounted() {
         this.setupPopover();
     },
 
@@ -116,49 +114,45 @@ export default {
 
             const options = {
                 animateFill: false,
-                // Default 'none' to 'fade', as it's not a valid Tippy.js option. The effect of no transition is achieved by `duration: 0` below.
+                // Use 'fade' when animation is 'none', as 'none' it's not a valid Tippy.js option.
+                // The effect of no transition is achieved by `duration: 0` below.
                 animation: this.animation === 'none' ? 'fade' : this.animation,
+                appendTo: this.appendToBody ? document.body : this.triggerEl.parentElement,
                 arrow: false,
                 content: this.$el,
-                delay: [this.openDelay, 0],
+                delay: [0, 0],
                 distance: 0,
-                duration: this.animation === 'none' ? 0 : 250,
-                hideOnClick: true,
+                duration: this.animation === 'none' ? 0 : [250, 200],
+                hideOnClick: false,
+                ignoreAttributes: true,
                 interactive: true,
-                // lazy: false,
+                lazy: true,
                 multiple: true,
-                offset: this.offset,
                 onHidden: this.onHidden,
                 onHide: this.onClose,
                 onShow: this.onOpen,
                 onShown: this.onShown,
-                performance: true,
                 placement: this.position,
-                shouldPopperHideOnBlur: () => false,
+                role: 'dialog',
                 theme: 'ui-popover',
-                trigger: this.openOn.indexOf('hover') === -1 ?
-                    this.openOn :
-                    this.openOn.replace('hover', 'mouseenter') // COMPAT: Support 'hover' for `openOn` prop,
+                trigger: this.openOn.replace('hover', 'mouseenter'),
+                popperOptions: {
+                    modifiers: {
+                        computeStyle: {
+                            // Disable GPU acceleration to fix blurry text in popover on Windows (Chrome)
+                            // https://github.com/twbs/bootstrap/issues/23590
+                            gpuAcceleration: !(window.devicePixelRatio < 1.5 && /Win/.test(navigator.platform))
+                        }
+                    }
+                }
             };
 
             if (!this.constrainToScrollParent) {
-                options.popperOptions = {
-                    modifiers: {
-                        preventOverflow: {
-                            enabled: false
-                        },
-                        hide: {
-                            enabled: false
-                        }
-                    }
-                };
+                options.popperOptions.modifiers.preventOverflow = { enabled: false };
+                options.popperOptions.modifiers.hide = { enabled: false };
             }
 
-            if (!this.appendToBody) {
-                options.appendTo = this.triggerEl.parentNode;
-            }
-
-            this.tip = tippy.one(this.triggerEl, options);
+            this.tip = tippy(this.triggerEl, options);
 
             if (this.disabled) {
                 this.tip.disable();
@@ -167,10 +161,14 @@ export default {
 
         destroyPopover() {
             if (this.tip) {
-                window.removeEventListener('scroll', this.onScroll);
+                this.removeCloseEventListeners();
                 this.tip.destroy();
                 this.tip = null;
             }
+        },
+
+        isOpen() {
+            return this.tip && this.tip.state.isVisible;
         },
 
         open() {
@@ -189,53 +187,90 @@ export default {
         toggle(options = { returnFocus: true }) {
             if (this.tip) {
                 this.returnFocus = options.returnFocus;
-                this.tip[this.tip.state.isVisible ? 'hide' : 'show']();
+                this.tip[this.isOpen() ? 'hide' : 'show']();
             }
         },
 
-        isOpen() {
-            return this.tip !== null && this.tip.state.isVisible;
+        scheduleUpdate() {
+            if (this.tip) {
+                this.tip.popperInstance.scheduleUpdate();
+            }
         },
 
         onOpen() {
-            if (this.closeOnScroll) {
-                document.addEventListener('scroll', this.onScroll, true);
-            }
+            this.addCloseEventListeners();
 
             classlist.add(this.triggerEl, 'has-dropdown-open');
 
             this.$emit('open');
         },
 
-        onShown() {
-            this.lastFocusedElement = document.activeElement;
-            this.$refs.focusContainer.focus();
-
-            this.$emit('reveal');
-        },
-
         onClose() {
-            if (this.closeOnScroll) {
-                document.removeEventListener('scroll', this.onScroll, true);
+            if (this.returnFocus && this.lastFocusedElement) {
+                this.lastFocusedElement.focus();
             }
+
+            this.removeCloseEventListeners();
 
             classlist.remove(this.triggerEl, 'has-dropdown-open');
 
             this.$emit('close');
-        },
 
-        onHidden() {
-            if (this.lastFocusedElement && this.returnFocus) {
-                this.lastFocusedElement.focus();
-            }
-
-            this.$emit('hide');
+            // Reset return focus
             this.returnFocus = true;
         },
 
-        onScroll(e) {
-            if (this.isOpen() && !this.$el.contains(e.target)) {
-                this.close();
+        onShown() {
+            this.lastFocusedElement = document.activeElement;
+            this.$refs.focusContainer.focus();
+            this.$emit('reveal');
+        },
+
+        onHidden() {
+            this.$emit('hide');
+        },
+
+        closeOnExternal(event, closeOptions) {
+            if (!this.$el.contains(event.target)) {
+                this.close(closeOptions);
+            }
+        },
+
+        addCloseEventListeners() {
+            this.removeCloseEventListeners();
+
+            // Add event listeners in the next tick, otherwise they're triggered immediately
+            setTimeout(() => {
+                this.removeExternalClickListener = events.on('click', document, e => {
+                    this.closeOnExternal(e, { returnFocus: false });
+                });
+
+                this.removeEscListener = events.onKeydown(27, document, () => {
+                    this.close({ returnFocus: true });
+                });
+
+                if (this.closeOnScroll) {
+                    this.removeScrollListener = events.on('scroll', document, e => {
+                        this.closeOnExternal(e, { returnFocus: true });
+                    });
+                }
+            }, 0);
+        },
+
+        removeCloseEventListeners() {
+            if (this.removeExternalClickListener) {
+                this.removeExternalClickListener();
+                this.removeExternalClickListener = null;
+            }
+
+            if (this.removeEscListener) {
+                this.removeEscListener();
+                this.removeEscListener = null;
+            }
+
+            if (this.removeScrollListener) {
+                this.removeScrollListener();
+                this.removeScrollListener = null;
             }
         }
     },